PRO-2995: extract the shape-declaration walk and redaction out of contract.rb
`contract.rb` grew from 1,268 to 2,439 lines on this branch, and the `expects`/`exposes` DSL a
contributor actually reads ended up buried in machinery that exists for long-tail hardening: the
declaration-time walk over a caller-supplied `shape:` graph, the `sensitive:` redaction derivation and
masking, and the option-container copy. Three moves, and nothing else:

* `Core::Contract::ShapeDeclaration` (new) — the declaration walk and snapshot plus its error helpers:
  `_validate_and_snapshot_shape!`, `_snapshot_declared_shape!`, `_walk_shape_graph!`,
  `_check_and_copy_shape_members!`, `_snapshot_member_attributes!`, the two `_symbol_keyed_member_*`
  canonicalizers, `_spend_paths!`, the seven `_raise_*` messages, `_describe_shape_member`,
  `_member_owner_label`, and the `WalkedShape`/`ShapeWalk`/`CYCLIC_SHAPE` private constants. All 17 stay
  private, declared by the same `private :…` list moved with them.
* `Core::Contract::Redaction` (new) — `_contract_redaction` and its `ContractRedaction` memo, the
  candidate/static/dynamic collectors, `_context_slice` and `_filter_tolerating_cycles`, the shape-path
  derivations (top-level and ambient), the masking walk, `_sensitive_member_names`, and
  `SENSITIVE_FILTERED_MASK`. `_context_slice` moves rather than delegating: every line of it is
  redaction, and a one-line forwarder would be new indirection for nothing.
* `Internal::ShapeGraph` — the option-container copy (`detach_option_containers!`,
  `detached_option_bag`, `detached_option_array`, `reject_defaulting_option_container!`,
  `describe_own_methods`), which is a value-level mechanism over caller-supplied containers rather than
  anything contract-specific, and joins the bound primitives it already consumes (`copy_entries`,
  `detached_dup`, `supplies_default?`). The three that no other layer calls are `private_class_method`.

Both new modules are `include`d into `Contract::ClassMethods` rather than extended separately onto the
action class, so `ClassMethods` remains the single thing `Contract.included` extends and every method
keeps its implicit receiver, its visibility, and its position in the singleton lookup chain (the two
modules sit in the slot `ClassMethods` already occupied, ahead of `AmbientContext::ClassMethods`).

`FieldConfig`/`ShapeConfig`/`FieldOptionality`/`ShapeBuilder` and the three `Contract.validate_*!`
grammars stay: they are the value objects and the single-sourced value grammars every layer enters
through (`ShapeValidator` calls `Contract.validate_user_facing!` by name), not walk machinery.

A PURE MOVE, verified by output rather than by reading: all 263 harness rows pass and are byte-identical
to bd44a5da apart from anonymous-class object ids; the effective public surface of a representative
action class (`methods(false)`, public/protected singleton and instance methods) is byte-identical, and
the private surface loses exactly the five methods that moved onto `Internal::ShapeGraph` — nothing
gained, nothing widened. Declaration, projection and redaction cost are unchanged within noise. 4206
examples / 0 failures / 1 pending on Ruby 3.2.5, 3.3.6 and 3.4.1; rubocop clean, spec_rubocop 49/0,
dummy app 330/0.

No CHANGELOG entry: nothing observable changed, and the five relocated private methods were introduced
on this same unreleased branch.
