+++ lib/axn/core/contract.rb
@@ -6,6 +6,8 @@ require "active_support/core_ext/enumerable"
 require "active_support/core_ext/module/delegation"
 require "active_support/core_ext/object/blank"
 
+require "axn/core/contract/redaction"
+require "axn/core/contract/shape_declaration"
 require "axn/core/validation/fields"
 require "axn/core/flow/handlers/invoker"
 require "axn/internal/shape_graph"
@@ -230,6 +232,14 @@ module Axn
       end
 
       module ClassMethods
+        # The long-tail declaration/redaction machinery, extracted so this file stays the
+        # `expects`/`exposes` DSL a contributor reads. Included into ClassMethods rather than extended
+        # separately onto the action class, so ClassMethods remains the single thing `Contract.included`
+        # extends and every method is reached with the same implicit receiver, at the same visibility,
+        # and in the same lookup position as when it lived in this file.
+        include ShapeDeclaration
+        include Redaction
+
         # rubocop:disable Metrics/ParameterLists
         def expects(
           *fields,
@@ -375,203 +385,6 @@ module Axn
           end
         end
 
-        # Everything redaction derives from the DECLARATION rather than from a call: the candidate config
-        # set, whether resolving `sensitive:` needs an instance at all, the static field-name filter, the
-        # static shape paths, and — per shape — which members a mask has to descend into.
-        #
-        # All of it is a pure function of the class, because a stored shape graph is a declaration-time
-        # SNAPSHOT — every node a Hash axn built, every member a `ShapeConfig` axn constructed from values read
-        # once (`_snapshot_member_attributes!`). That is what makes an identity-keyed table sound at all: a
-        # retained member would have left `sensitive:` free to change with nothing the key could see, so the
-        # first read of a contract decided it and a later flip changed nothing. Re-deriving it per logged call
-        # instead made the cost of redaction scale with the size of that graph, and worst for the contracts
-        # that use no `sensitive:` at all: concluding "there is nothing to redact" is exactly what requires
-        # visiting every member, so a 24,000-member shape with no `sensitive:` anywhere cost more per logged
-        # call (~34ms) than the same shape with one.
-        #
-        # Keyed by the IDENTITY of the three config arrays, exactly as `_resolved_subfields` is and as
-        # `PropertyNames.validate_outbound!`'s verdict is: the arrays are copy-on-write, so ANY later
-        # declaration — a reopened class, a subclass, `Mountable`'s builder clearing them, a downstream gem
-        # assigning the attribute directly — mints new ones and the stale table misses on `equal?`. That is
-        # invalidation with no hook to keep in sync, which matters more here than for a schema cache: a
-        # stale answer does not merely disagree with the contract, it fails to redact a `sensitive:` field
-        # declared after the first logged call.
-        def _contract_redaction
-          internals = internal_field_configs
-          externals = external_field_configs
-          subfields = subfield_configs
-          memo = @_axn_contract_redaction
-          return memo if memo&.current?(internals, externals, subfields)
-
-          @_axn_contract_redaction = ContractRedaction.new(internals:, externals:, subfields:)
-        end
-
-        # The table `_contract_redaction` hands out. Mutable (so not a `Data`) because each fact is derived
-        # on first demand and written once, with `nil` meaning "not yet derived" — which none of the values
-        # can be. Unlocked deliberately: every slot's value is the one answer that contract will ever have, so
-        # a racing second derivation recomputes it and writes the same thing (the resolved-subfield cache's
-        # single-ivar-write reasoning). The two per-shape tables are the same bargain one level down — a lost
-        # race costs a repeat derivation, and a single Hash insert is not interruptible by another Ruby thread.
-        class ContractRedaction
-          attr_reader :internals, :externals, :subfields
-          attr_accessor :candidates, :dynamic, :static_fields, :filter,
-                        :static_shape_paths, :static_ambient_shape_paths
-
-          def initialize(internals:, externals:, subfields:)
-            @internals = internals
-            @externals = externals
-            @subfields = subfields
-            # Per shape / per config, by identity — neither defines a `hash`/`eql?` axn would want to run,
-            # and identity is the right question anyway: the stored graph is the one axn snapshotted.
-            @nested_members = {}.compare_by_identity
-            @member_names = {}.compare_by_identity
-          end
-
-          def current?(internals, externals, subfields)
-            @internals.equal?(internals) && @externals.equal?(externals) && @subfields.equal?(subfields)
-          end
-
-          def nested_members_for(shape, &derive) = @nested_members.fetch(shape) { @nested_members[shape] = derive.call }
-
-          def member_names_for(config, &derive) = @member_names.fetch(config) { @member_names[config] = derive.call }
-        end
-        private_constant :ContractRedaction
-
-        def inspection_filter
-          memo = _contract_redaction
-          memo.filter ||= ActiveSupport::ParameterFilter.new(sensitive_fields)
-        end
-
-        def sensitive_fields
-          _static_sensitive_fields
-        end
-
-        # Every config whose `sensitive:` participates in redaction: the declared inbound/outbound fields
-        # and subfields, plus (recursively) the members of any shape block they carry. Shape members live
-        # in validations[:shape][:members] at every depth, so the walk is uniform — a sensitive member at
-        # any nesting level contributes its name to the ParameterFilter set (which redacts by key name at
-        # any depth, array elements included). Single-sources the traversal for all three collectors.
-        # Derived from the arrays the memo is KEYED on, never re-read from the class: a table that answered
-        # from a newer contract than the one it is keyed to would be a stale answer wearing a valid key.
-        def _sensitive_candidate_configs
-          memo = _contract_redaction
-          memo.candidates ||= (memo.internals + memo.externals + memo.subfields)
-                              .flat_map { |config| _flatten_sensitive_candidates(config) }
-        end
-
-        def _flatten_sensitive_candidates(config, seen = nil, depth = 0)
-          # Through the shared seam: a member list that hides itself from `flat_map` would drop a
-          # `sensitive:` member from the redaction set, which leaks rather than merely disagreeing.
-          shape = Internal::ShapeGraph.shape_in(config.validations)
-          # Bounded BOTH ways, because the configs this walks need not have been DECLARED: the declaration walk
-          # rejects an untraversable graph and snapshots what it accepts, but `internal_field_configs` and friends
-          # are writable, so a config assigned onto a class carries whatever shape its author built — pointed back
-          # at itself (which `CycleGuard` sees) or minting a fresh nested shape on every read (which nothing sees
-          # but depth). This runs while a log line or an exception report is being built — once per contract, or
-          # per logged call for a `sensitive:` that resolves against the action — and the alternative is
-          # SystemStackError from a log line, a side channel taking down the call it observes.
-          #
-          # Nothing is lost by stopping at a cycle: it re-reaches members an enclosing frame is already
-          # collecting. Past the depth bound there IS no honest answer, and the wholesale mask does the work —
-          # `_shape_has_sensitive_member?` answers true there, so the whole value is redacted rather than
-          # filtered per member.
-          return [config] if nil.equal?(shape) || depth > Internal::ShapeGraph::MAX_NESTING
-
-          nested = Axn::Internal::CycleGuard.guard(shape, seen, on_cycle: []) do |open|
-            Internal::ShapeGraph.members(shape).flat_map { |member| _flatten_sensitive_candidates(member, open, depth + 1) }
-          end
-          [config, *nested]
-        end
-
-        def _static_sensitive_fields
-          memo = _contract_redaction
-          memo.static_fields ||= _sensitive_candidate_configs
-                                 .select { |c| _config_sensitive(c) == true }
-                                 .flat_map { |c| _sensitive_field_keys(c) }
-        end
-
-        # Whether resolving this contract's `sensitive:` needs an action instance — and therefore the ONE
-        # question every reuse decision below turns on. It can be one question because the value space is closed
-        # at declaration (see Contract.validate_sensitive!): over `true`/`false`/`nil`/Symbol/Proc, "carries a
-        # Proc or Symbol" and "would resolve differently with an instance than without" are the same set. A
-        # second, stricter predicate existed here to cover a truthy value that is neither — which the
-        # instanceless path counted as not-sensitive while the per-instance path truthiness-tested it as
-        # sensitive, observably so for a non-Hash shaped value — and rejecting that value at declaration is what
-        # removes the divergence instead of guarding it.
-        #
-        # `false` is one of the two answers, so the memo is guarded on `nil` rather than written with `||=`: an
-        # `||=` here re-ran the whole candidate walk on every logged call for precisely the contracts that have
-        # no dynamic `sensitive:` — which is nearly all of them.
-        def _has_dynamic_sensitive_fields?
-          memo = _contract_redaction
-          return memo.dynamic unless memo.dynamic.nil?
-
-          memo.dynamic = _sensitive_candidate_configs.any? do |config|
-            # `case`/`when` consults the real class, so a caller-supplied member's value cannot decide whether
-            # it needs an instance — the same non-dispatching test the declaration guard uses.
-            case _config_sensitive(config)
-            when ::Proc, ::Symbol then true
-            else false
-            end
-          end
-        end
-
-        def _resolve_sensitive_fields(action_instance)
-          return _static_sensitive_fields unless _has_dynamic_sensitive_fields?
-
-          _sensitive_candidate_configs
-            .select { |config| _resolve_sensitive_value(_config_sensitive(config), action_instance) }
-            .flat_map { |c| _sensitive_field_keys(c) }
-        end
-
-        # A shape member's contract is duck-typed — ShapeValidator requires only #field/#validations,
-        # and `shape: { members: [...] }` may be supplied raw with member objects that implement no
-        # more than that. `#sensitive` is optional (absent on such a raw member), so read it defensively
-        # and treat a missing reader as `false`, mirroring how the validator treats a missing
-        # #method_call as not opted in. Read through Internal::ShapeGraph, so a member that DEFINES
-        # `sensitive:` cannot opt out of redaction by denying the reader.
-        #
-        # That costs a Method allocation per read (~0.25µs, against ~0.04µs for `respond_to?` plus a
-        # dispatch) and is spent deliberately: the reads here are per config on a derivation, and a
-        # member hiding from redaction leaks a secret. `_sensitive_shape_paths` makes the opposite call for
-        # the opposite reason — it reads axn's OWN configs, which cannot lie, so it takes the cheap
-        # `shape_in` path. The tradeoff differs because the exposure does.
-        def _config_sensitive(config)
-          Internal::ShapeGraph.read(config, :sensitive) || false
-        end
-
-        # A sensitive `model:` field also redacts its generated `<field>_id` alias (the id is as
-        # sensitive as the record). Non-model fields contribute only their own key.
-        def _sensitive_field_keys(config)
-          keys = [config.field]
-          keys << Internal::FieldConfig.model_id_key(config.field) if config.validations[:model]
-          keys
-        end
-
-        # The first three arms are the whole declared grammar (see Contract.validate_sensitive!); `!!` is what
-        # turns a predicate's arbitrary return value into a decision. The `else` is now reachable only by `nil`
-        # — which `_config_sensitive` already normalizes to `false` on every path that reads a config — and is
-        # kept fail-SAFE rather than dropped: were an out-of-grammar value ever to arrive here, redacting a
-        # truthy one is the direction that cannot leak. This is the branch whose truthiness test used to
-        # disagree with the instanceless path about `sensitive: "yes"`, and closing the value space is what
-        # removed the disagreement rather than a second predicate guarding it.
-        def _resolve_sensitive_value(sensitive, action_instance)
-          case sensitive
-          when true, false
-            sensitive
-          when Symbol
-            !!action_instance.send(sensitive)
-          when Proc
-            !!action_instance.instance_exec(&sensitive)
-          else
-            !!sensitive
-          end
-        end
-
-        def _build_instance_filter(action_instance)
-          ActiveSupport::ParameterFilter.new(_resolve_sensitive_fields(action_instance))
-        end
-
         def _declared_fields(direction)
           raise ArgumentError, "Invalid direction: #{direction}" unless direction.nil? || %i[inbound outbound].include?(direction)
 
@@ -583,190 +396,6 @@ module Axn
 
           configs.map(&:field)
         end
-
-        # Internal method for filtering context data by direction
-        # Used by instance methods (inputs_for_logging, outputs_for_logging) and async exception reporting
-        # When action_instance is provided, dynamic sensitive fields are resolved against that instance.
-        def _context_slice(data:, direction: nil, action_instance: nil)
-          filter = if action_instance && _has_dynamic_sensitive_fields?
-                     _build_instance_filter(action_instance)
-                   else
-                     inspection_filter
-                   end
-          sliced = _mask_unfilterable_shapes(data.slice(*_declared_fields(direction)), _sensitive_shape_paths(action_instance), action_instance)
-          _filter_tolerating_cycles(filter, sliced)
-        end
-
-        # ActiveSupport::ParameterFilter has no cycle guard of its own, so a self-referential value blows
-        # its stack — and since a filter is active for any action declaring `sensitive:`, that cost the
-        # WHOLE slice: both auto_log lines dropped to warnings and the exception report degraded to empty
-        # inputs/outputs, losing the fields that had nothing to do with the cycle.
-        #
-        # Retried on a decycled copy rather than decycling up front, so acyclic data (all of it, in
-        # practice) pays nothing: no extra walk, no copy. The retry can only be reached by data that
-        # already failed, where one wasted filter pass is irrelevant.
-        def _filter_tolerating_cycles(filter, data)
-          filter.filter(data)
-        rescue SystemStackError
-          filter.filter(Axn::Internal::CycleGuard.decycle(data))
-        end
-
-        # Per-element `sensitive:` redaction works by adding the member's key name to an
-        # `ActiveSupport::ParameterFilter`, which redacts Hash keys at any depth — so a member inside a
-        # Hash (or an Array of Hashes) is filtered precisely, siblings preserved. But the filter only
-        # descends into Hashes: an object-backed shape value (a Data/Struct/PORO), or a malformed non-Hash
-        # value where a Hash was expected (which reaches logging before inbound validation can reject it),
-        # is opaque to it and would print whole. So for every field/subfield whose shape carries a
-        # sensitive member, this walks to the shaped value and replaces a non-Hash value in a
-        # member-bearing position with the mask — over-redacting the whole value (its non-sensitive
-        # siblings included) rather than risk leaking the secret. Applied to logs, exception context,
-        # and `inspect`.
-        def _mask_unfilterable_shapes(data, shape_paths, action_instance)
-          return data unless data.is_a?(Hash)
-
-          shape_paths.reduce(data) do |acc, (wire_path, shape)|
-            _mask_value_at_path(acc, wire_path, shape, action_instance)
-          end
-        end
-
-        # Single-field entry — inspect renders one field at a time. Reuses the whole-hash pass on a
-        # one-key hash so a subfield shape rooted under `field` is masked at its nested position too.
-        def _mask_unfilterable_shape_value(field, value, action_instance)
-          _mask_unfilterable_shapes({ field => value }, _sensitive_shape_paths(action_instance), action_instance)[field]
-        end
-
-        # `[(wire_path, shape)]` for every field/subfield whose shape carries a sensitive member. A
-        # top-level field's path is `[field]`; a subfield's is its resolved wire path (from the
-        # SubfieldTree cache), so a shape declared on a subfield — `expects :person, on: :payload, …
-        # do … end` — is masked at `payload[:person]`, not just where a top-level shape lives.
-        # Memoized unless a `sensitive:` resolves against the action (`_has_dynamic_sensitive_fields?`), which
-        # is what keeps a logged call from paying for the whole stored graph. A contract that DOES need the
-        # instance still derives per call — correctness requires it, and a memo ignoring the instance would
-        # over-redact a `sensitive: :flag` member whose flag is false, which looks safe and is a behavior
-        # change.
-        def _sensitive_shape_paths(action_instance)
-          return _derive_sensitive_shape_paths(action_instance) if _has_dynamic_sensitive_fields?
-
-          memo = _contract_redaction
-          memo.static_shape_paths ||= _derive_sensitive_shape_paths(nil)
-        end
-
-        # Over the arrays the memo is keyed on, for the reason `_sensitive_candidate_configs` gives.
-        def _derive_sensitive_shape_paths(action_instance)
-          memo = _contract_redaction
-          (memo.internals + memo.externals + memo.subfields).filter_map do |config|
-            shape = Internal::ShapeGraph.shape_in(config.validations)
-            next unless shape && _shape_has_sensitive_member?(shape, action_instance)
-
-            wire_path = config.subfield? ? _resolved_subfields.index[config]&.wire_path : [config.field]
-            next unless wire_path
-
-            [wire_path, shape]
-          end
-        end
-
-        # The ambient analog of `_sensitive_shape_paths`: `[(wire_path_within_ambient, shape)]` for
-        # every ambient subfield whose shape carries a sensitive member. Ambient shapes are leaf nodes
-        # (`_check_ambient_shape_placement!`), so their value is copied whole and may be a non-Hash the
-        # ParameterFilter can't descend into — this feeds `_mask_unfilterable_shapes` to mask it. The
-        # ambient tree's wire paths are rooted at the synthetic `:ambient_context` segment; drop it,
-        # since the mask applies to the ambient VALUE the reader returns, not a hash wrapped under an
-        # `:ambient_context` key.
-        def _sensitive_ambient_shape_paths(action_instance)
-          return _derive_sensitive_ambient_shape_paths(action_instance) if _has_dynamic_sensitive_fields?
-
-          memo = _contract_redaction
-          memo.static_ambient_shape_paths ||= _derive_sensitive_ambient_shape_paths(nil)
-        end
-
-        def _derive_sensitive_ambient_shape_paths(action_instance)
-          _ambient_subfield_tree.index.filter_map do |config, path|
-            shape = Internal::ShapeGraph.shape_in(config.validations)
-            next unless shape && _shape_has_sensitive_member?(shape, action_instance)
-
-            [path.wire_path.drop(1), shape]
-          end
-        end
-
-        # Walk `wire_path` through `value` — Hash keys in either symbol or string form (extraction
-        # accepts both), mapping across arrays — and mask the shaped value at the leaf. Every present
-        # key form is masked (see `_present_key_variants`); an absent key is left alone. A non-Hash,
-        # non-Array intermediate with path still remaining is an object-backed parent (a `method_call:`
-        # subfield reads the sensitive shape off it) that ParameterFilter can't descend into — mask it
-        # wholesale rather than leak the sensitive member nested inside; a nil/scalar intermediate is
-        # preserved (nothing to reach or leak).
-        # The Hash branch consumes a path segment per level so it always terminates; the Array branch
-        # maps with the path UNCHANGED, so a self-referential array would recurse until the stack
-        # blows. A revisited array masks wholesale (not the `[...]` placeholder, which would be a
-        # placeholder String masquerading as data) — the same over-redact-rather-than-leak call as
-        # `_mask_opaque_or_preserve`, since we cannot descend to redact the sensitive member inside.
-        def _mask_value_at_path(value, wire_path, shape, action_instance, seen = nil)
-          return _mask_shape_value(value, shape, action_instance) if wire_path.empty?
-
-          if value.is_a?(Array)
-            return Axn::Internal::CycleGuard.guard(value, seen, on_cycle: SENSITIVE_FILTERED_MASK) do |nested|
-              value.map { |element| _mask_value_at_path(element, wire_path, shape, action_instance, nested) }
-            end
-          end
-          return _mask_opaque_or_preserve(value) unless value.is_a?(Hash)
-
-          _present_key_variants(value, wire_path.first).reduce(value) do |acc, key|
-            acc.merge(key => _mask_value_at_path(acc[key], wire_path.drop(1), shape, action_instance, seen))
-          end
-        end
-
-        # A non-Hash/Array value in a member-bearing position. `nil` is preserved — it is valid absent
-        # data (a nil-tolerant shape) with nothing to leak, and masking it would make absent data look
-        # redacted. Anything else is malformed for a shaped field (which expects a Hash/object with
-        # members) and ParameterFilter can't redact into it: a structured object could expose the member
-        # via `inspect`, and a bare scalar could itself BE the sensitive value the caller mis-supplied
-        # (`items: ["111-11-1111"]`). Both reach logging before validation rejects them, so mask.
-        def _mask_opaque_or_preserve(value)
-          value.nil? ? value : SENSITIVE_FILTERED_MASK
-        end
-
-        # Every form of `key` present in `hash` — the key as-is, its string form, and its symbol form.
-        # Extraction accepts symbol and string keys (reading symbol-first) and a member/wire-path name
-        # may be declared in either form, so a single logical key can appear under more than one form in
-        # the same Hash; mask them all, since every form is logged and any could hold the secret.
-        def _present_key_variants(hash, key)
-          [key, key.to_s, key.to_s.to_sym].uniq.select { |variant| hash.key?(variant) }
-        end
-
-        # Whether a shape tree carries a `sensitive:` member anywhere (direct, or in a nested shape).
-        # A nil action_instance (async reporting, no instance to resolve a dynamic predicate against)
-        # counts only static `sensitive: true`, matching the static `inspection_filter` used there.
-        # Bounded for the same reasons as `_flatten_sensitive_candidates`. The cycle answer is exact rather than
-        # defensive: a cyclic branch re-reaches the very members the enclosing frame is already testing, so
-        # answering `false` for it decides nothing on its own and no `sensitive:` member can hide in one. The
-        # depth answer cannot be exact, so it is fail-safe instead (see below).
-        def _shape_has_sensitive_member?(shape, action_instance, seen = nil, depth = 0)
-          # Past the depth bound the answer is TRUE, not false: a graph that deep is one minting fresh nested
-          # shapes on every read, so nothing can enumerate what is inside it — and the fail-safe answer for
-          # redaction is "assume a secret", which masks the value wholesale rather than logging it in the clear.
-          # Unreachable for a declared graph, since the declaration walk rejects one this deep.
-          return true if depth > Internal::ShapeGraph::MAX_NESTING
-
-          Axn::Internal::CycleGuard.guard(shape, seen, on_cycle: false) do |open|
-            Internal::ShapeGraph.members(shape).any? do |member|
-              _member_sensitive?(member, action_instance) ||
-                (_member_shape(member) && _shape_has_sensitive_member?(_member_shape(member), action_instance, open, depth + 1))
-            end
-          end
-        end
-
-        def _member_sensitive?(member, action_instance)
-          sensitive = _config_sensitive(member)
-          return sensitive == true if action_instance.nil?
-
-          _resolve_sensitive_value(sensitive, action_instance)
-        end
-
-        # The shape a config or member carries, or nil when it carries none — read without dispatching
-        # anything a raw `shape:` kwarg's objects can define (see Internal::ShapeGraph), so a member
-        # lying about its type or its readers cannot hide a nested shape from a walk that reflection,
-        # validation and redaction all still descend into.
-        def _member_shape(member) = Internal::ShapeGraph.nested_shape(member)
 
         # Reject `user_facing:` on any member of an `exposes` shape, at any depth. The block form
         # catches it in `_build_shape_member` on key presence; a raw `shape:` kwarg supplies pre-built
@@ -790,534 +419,6 @@ module Axn
             _reject_outbound_shape_user_facing!(_member_shape(member))
           end
         end
-
-        # How a message names a shape member: its declared name when it has one, else its class. Never
-        # its `inspect` — that is the member's own code running while the failure is being reported, and
-        # an exception from it would replace the declaration error (escaping every rescue when it is
-        # outside StandardError). A member with no `#field` has no name to print, and its class is the
-        # only thing left that identifies it.
-        #
-        # `name` is passed in rather than read here, and every caller reads it on the way DOWN, before it
-        # has a failure to report. Re-reading it here would be the same hazard as `inspect` one step
-        # removed: a reader that raises (or answers differently) on a second read replaces the declaration
-        # error with the caller's exception — verified against a member whose `field` raises the second time
-        # it is called.
-        def _describe_shape_member(member, name)
-          return "of class #{Axn::Internal::ClassName.of(member)}" if Internal::ShapeGraph.missing?(name)
-
-          "`#{_shape_member_label(name)}`"
-        end
-
-        # A shape member's name is an object property in the reflected schema on exactly the same terms as a
-        # field's, so it carries the same promise. Walks RESOLVED members rather than checking inside
-        # ShapeBuilder because the `do…end` form routes through `_build_shape_member` but a raw `shape:`
-        # kwarg supplies pre-built members that never do — the same reason
-        # `_reject_outbound_shape_user_facing!` walks. Recursion covers a member's own nested block.
-        #
-        # A member that answers to no `#field` is rejected rather than skipped. The documented member contract
-        # is `#field` PLUS `#validations`, and runtime validation reads `member.field` for every member it
-        # validates — so a nameless member declared cleanly, reflected as nothing at all, and then raised
-        # NoMethodError on the first call. Skipping it in the guard while the consumer dispatches it anyway is
-        # the guard/consumer divergence this walk exists to eliminate. What the tolerance was ever for is the
-        # opposite direction: a member that DEFINES `field` cannot escape the check by denying the reader,
-        # decided from the real method table (see Internal::ShapeGraph), because reflection reads that name
-        # regardless and the two must agree.
-        #
-        # This is the first shape walk on every declaration path, so it is where an untraversable graph is
-        # rejected on behalf of all of them: a graph reaching validation, reflection or redaction has
-        # already been rejected here if it is untraversable AS READ. What no declaration-time walk can
-        # promise is a graph that answers a LATER read differently — a `[]` or a reader giving two answers
-        # is a caller contradicting itself, not a guard missing something (see Internal::ShapeGraph).
-        #
-        # `via`/`via_name` name the member whose nested shape is being entered — the name captured on the
-        # way down, never re-read while a failure is being reported.
-        #
-        # Two ways a graph can be untraversable, and each needs its own answer. A graph that CONTAINS
-        # itself repeats an object, which `CycleGuard` detects by identity. A graph that GENERATES itself —
-        # a shape Hash whose `[]` builds a fresh nested shape on every read — never repeats an object, so
-        # no identity guard can see it; it is infinitely deep rather than cyclic, and only a depth cap
-        # stops it. Both otherwise end in SystemStackError, outside StandardError, escaping every rescue.
-        # THE declaration walk over a caller-supplied shape graph, and the only one: it rejects what cannot be
-        # declared and copies what can, in a single pass. Returns the copy, which the caller stores as the
-        # declared shape.
-        #
-        # Fused rather than a check pass followed by a copy pass, because the two have to agree about what the
-        # members ARE: a list that answers a second `each` differently would leave the class holding members no
-        # check ever saw.
-        #
-        # It also bounds the graph's SIZE, in member paths (see ShapeGraph::MAX_MEMBER_PATHS) — deliberately not
-        # in emitted JSON properties, which is a different limit belonging to reflection, derived from what the
-        # emitter emits, and applied at projection. This one is about the graph itself: the walks that read a
-        # stored graph on a live call have no per-reference memo, so a graph that multiplies out costs the CALL.
-        # Counting it here is what keeps that knowable at declaration, where the author is present.
-        #
-        # So a caller's members list is read exactly ONCE per declaration, and the answer it gave is the
-        # contract: a list that would answer a later read differently never gets that read. What it cannot do is
-        # decline the first one — the declaration is not knowable without it — so a list that raises on being
-        # read raises at declaration, which is the intended outcome and the right one: the author is standing
-        # there, rather than the failure landing on whoever first reflects the class.
-        def _validate_and_snapshot_shape!(shape, fields)
-          _walk_shape_graph!(shape, nil, [Internal::ShapeGraph::MAX_MEMBER_PATHS, fields]).copy
-        end
-
-        # Stores the copy in place of the caller's shape, and only when there is one to copy: a field that
-        # declared no `shape:` must not gain the key here, and a `shape:` that is not a Hash is left exactly as
-        # it came for the container check to reject.
-        def _snapshot_declared_shape!(validations, fields)
-          shape = Internal::ShapeGraph.hash_or_nil(validations[:shape])
-          return if nil.equal?(shape)
-
-          validations[:shape] = _validate_and_snapshot_shape!(shape, fields)
-        end
-
-        # What one walked shape yields. The path count travels with the copy because a shape REUSED by two
-        # members is walked (and copied) once but COUNTED twice: sharing is exactly how a graph multiplies out,
-        # so the second reference charges the whole total its subtree expands to.
-        WalkedShape = Data.define(:copy, :paths)
-        private_constant :WalkedShape
-
-        # The remaining allowance, and the fields to name if it runs out — a two-element Array rather than an
-        # object, because it is threaded through every level of one walk and the label is only ever built on the
-        # failure path.
-        def _spend_paths!(allowance, paths)
-          allowance[0] -= paths
-          return unless allowance[0].negative?
-
-          _raise_too_many_member_paths!(allowance[1])
-        end
-
-        def _raise_too_many_member_paths!(fields)
-          raise ArgumentError,
-                "the shape on #{_inspect_field_name(fields.first)} has more than " \
-                "#{Internal::ShapeGraph::MAX_MEMBER_PATHS} member paths — a nested shape object reused by " \
-                "sibling members multiplies out, so N levels of two-way sharing are 2^N distinct paths, and " \
-                "every walk of the stored graph pays one step per path: runtime validation walks it on each " \
-                "call, and redaction re-walks it per logged call whenever a `sensitive:` resolves against the " \
-                "action (measured: 786,000 paths cost about 1.3 seconds per log line, and about two seconds " \
-                "for the one derivation any contract makes on its first). Give each member its own nested " \
-                "shape, or flatten the nesting. This is a bound on the graph, not on what a schema emits — an " \
-                "oversized SCHEMA is reported separately, when a projection is first built."
-        end
-
-        def _walk_shape_graph!(shape, walk, allowance, via: nil, via_name: nil)
-          hash = Internal::ShapeGraph.hash_or_nil(shape)
-          # Not a shape, so it has no members to walk and nothing to copy — returned as it came, for the
-          # container check downstream to reject.
-          return WalkedShape.new(copy: shape, paths: 0) if nil.equal?(hash)
-
-          # Ahead of "does it supply members?", because a shape whose `members:` comes from a default supplies
-          # them to this walk and not to the snapshot it produces — the reverse mismatch of the one below, and
-          # reported as the defect it is rather than as a shape with no members.
-          _reject_defaulting_option_container!(hash) do
-            nil.equal?(via) ? "the `shape:`" : "the nested `shape:` at shape member #{_describe_shape_member(via, via_name)}"
-          end
-
-          walk ||= ShapeWalk.new(seen: nil, walked: {}.compare_by_identity, depth: 0)
-          walked = walk.walked[hash]
-          unless nil.equal?(walked)
-            _spend_paths!(allowance, walked.paths)
-            return walked
-          end
-
-          _raise_shape_too_deep!(via, via_name) if walk.depth > Internal::ShapeGraph::MAX_NESTING
-
-          # Read as SUPPLIED, so a shape that names no members is told apart from one naming an empty list.
-          members = Internal::ShapeGraph.declared_members(hash)
-          _raise_missing_shape_members!(via, via_name) if nil.equal?(members)
-
-          walked = Axn::Internal::CycleGuard.guard(hash, walk.seen, on_cycle: CYCLIC_SHAPE) do |nested|
-            _check_and_copy_shape_members!(hash, members, walk.with(seen: nested), allowance)
-          end
-          _raise_cyclic_shape!(via, via_name) if CYCLIC_SHAPE.equal?(walked)
-
-          walk.walked[hash] = walked
-          walked
-        end
-
-        # The state one walk carries. `seen` is the ANCESTRY set `CycleGuard` pushes and pops, which is what
-        # makes a diamond (one nested shape reused by two siblings) legal rather than a false cycle.
-        # `walked` is the complement: shapes already walked to completion, which never need walking again —
-        # without it, that same legal diamond costs 2^depth walks (measured: 18 levels took 1.4s, 22 took
-        # 22s), so a generated-but-honest schema with shared sub-shapes hung at class definition. Keyed by
-        # identity, per declaration, and populated only AFTER a shape has passed, so a memoized entry always
-        # means "already verified" — and carries that shape's copy, so a shape reused by two members is read
-        # from the caller once and both members store the one copy. A shape that answers a later read
-        # differently is the inconsistent-reader limit above, not something re-walking would have caught.
-        ShapeWalk = Data.define(:seen, :walked, :depth)
-        private_constant :ShapeWalk
-
-        # Sentinel for "this shape was already open on the path" — a private object rather than a value a
-        # declaration could produce, so nothing a caller supplies can be mistaken for it, and identity is
-        # asked of the sentinel so no caller's `equal?` is dispatched.
-        CYCLIC_SHAPE = Object.new.freeze
-        private_constant :CYCLIC_SHAPE
-
-        # One node: its members checked, counted, and copied. Each member's name is read exactly ONCE, into a
-        # (member, name) pair, and every later use of it — a collision message, a cycle or depth report against
-        # a nested shape, the copy stored for it — reads that capture rather than the member again. Same reasoning
-        # as the renderer's one-#to_s-per-key rule: a second read may disagree with the first or raise something
-        # that is not even a StandardError, replacing the diagnosis with the escape these guards exist to prevent.
-        #
-        # Reading the name once is not enough on its own, because CANONICALIZING it is a second dispatch on the
-        # same caller object: a String subclass whose `to_sym` answers `:alpha` and then `:collide` gave the
-        # duplicate check one property name and `ShapeConfig`'s constructor another, so two declared members were
-        # stored under one property, `member_properties` emitted one, and nothing raised. So the canonical Symbol
-        # is computed once too, beside the check that judges it, and threaded to the snapshot — which is what
-        # makes "the guard judged the property this member is stored under" true rather than probable.
-        def _check_and_copy_shape_members!(hash, members, walk, allowance)
-          named = members.map { |member| [member, Internal::ShapeGraph.fetch(member, :field)] }
-          named.each { |member, name| _raise_nameless_member!(member, name) if Internal::ShapeGraph.missing?(name) }
-          names = named.map { |_member, name| name }
-          # Ahead of every other name check: a name that is not a String or a Symbol has two independent
-          # renderings rather than one property (see Contract.validate_shape_member_name!), so asking whether
-          # it is renderable, or whether it collides, would be asking about only one of them. A ShapeConfig
-          # was already held to this in its constructor; a duck-typed member reaches it only here.
-          names.each { |name| Contract.validate_shape_member_name!(name) }
-
-          # Only the SAME key declared twice in one block is judged here — keyed by `to_sym`, which is exactly
-          # what `Schema#member_properties` keys a member's property by, so the two agree about what "the same
-          # member" is. Two members whose names merely CANONICALIZE alike are two distinct keys at one node,
-          # which is a property collapse rather than a repeat declaration, and the one claim space judges every
-          # such collapse (`_reject_colliding_property_claims!`) so no mechanism pair can slip between checks.
-          # This one cannot move there: two identical claims are a legal MERGE by that rule, while two
-          # identical members of one block are a genuine duplicate.
-          #
-          # The key each member is judged under is CAPTURED here (into a (member, name, key) triple) and stored
-          # as that member's `field`, rather than converted again downstream — see the note above. Computed
-          # inside this loop rather than in a pass of its own so the order of failures is unchanged: a name whose
-          # `to_sym` raises (bytes with no Symbol, say) is still reached after every earlier member has been
-          # judged, so a duplicate declared ahead of it is still reported as the duplicate it is.
-          claimed = {}
-          keyed = named.map do |member, name|
-            key = name.to_sym
-            _raise_duplicate_member!(name) if claimed.key?(key)
-
-            claimed[key] = name
-            [member, name, key]
-          end
-
-          child = walk.with(depth: walk.depth + 1)
-          paths = 0
-          copied = keyed.map do |member, name, key|
-            # Charged BEFORE this member is snapshotted, and before its nested shape is walked, so a graph that
-            # multiplies out is rejected while the work done on it is still bounded by the allowance.
-            _spend_paths!(allowance, 1)
-            paths += 1
-            # `validations` is read ONCE and threaded to every use — the nested shape to walk, and the snapshot
-            # of this member. A second read is a second answer the caller can give.
-            validations = _symbol_keyed_member_validations(member, name)
-            _raise_member_without_validations!(member, name) if nil.equal?(validations)
-            # Every other attribute read (and grammar-checked) BEFORE the nested walk, so a member carrying both
-            # a bad `sensitive:` and an untraversable nested shape is reported as the value defect it is.
-            attributes = _snapshot_member_attributes!(member, name, key, validations)
-            nested = Internal::ShapeGraph.hash_or_nil(validations[:shape])
-            unless nil.equal?(nested)
-              inner = _walk_shape_graph!(nested, child, allowance, via: member, via_name: name)
-              paths += inner.paths
-              validations[:shape] = inner.copy
-            end
-            ShapeConfig.new(**attributes)
-          end
-
-          WalkedShape.new(copy: Internal::ShapeGraph.snapshot_node(hash, copied), paths:)
-        end
-
-        # Everything a stored member carries, read off the caller's object exactly ONCE and held to its grammar
-        # on the way — so what the class stores is axn's own `ShapeConfig` and the caller keeps nothing live in a
-        # declared contract. Read through `ShapeGraph`, so a member that denies a reader it defines is still held
-        # to the rules, and the value snapshotted is the one that reader gives.
-        #
-        # Snapshotting rather than storing the object is what makes "the contract is what you declared" true for
-        # a member too: a retained member's `sensitive:` could be flipped after the class was declared (changing
-        # what redaction masked, or not, depending only on whether anything had asked yet), its canonicalized
-        # `validations` were computed and then thrown away (so a String-keyed `type:` bag validated nothing and
-        # reflected nothing), its option containers stayed aliased where a field's were detached, and the nested
-        # shape it carried could gain members or be pointed at itself afterwards.
-        #
-        # The grammars for `sensitive:`/`user_facing:` are checked HERE, ahead of `ShapeConfig`'s constructor,
-        # only so the error keeps its place in this walk's order; the constructor enforces the same rules for the
-        # block form, where it is the first thing to see them. A falsey `user_facing:` is "not opted in" and has
-        # no grammar to meet — `nil` is what an absent reader answers.
-        #
-        # `description` is read directly rather than taken from `metadata`, because that is where reflection
-        # reads it from (`Schema.declared_attribute`) and a duck-typed member may define the reader without any
-        # metadata at all; folding it into the metadata Hash is what `ShapeConfig#description` then answers with.
-        #
-        # `name` and `key` are the same name in its two roles: `name` is what a message about this member RENDERS
-        # (a raw String keeps the spelling the caller wrote, and an unrenderable one its escaped form), while
-        # `key` is the canonical Symbol the duplicate check already judged — and so the property this member is
-        # stored, validated and emitted under. Both come from the one read; nothing here converts either again.
-        def _snapshot_member_attributes!(member, name, key, validations)
-          sensitive = Internal::ShapeGraph.read(member, :sensitive)
-          Contract.validate_sensitive!(sensitive)
-          user_facing = Internal::ShapeGraph.read(member, :user_facing)
-          Contract.validate_user_facing!(user_facing) if user_facing
-
-          metadata = _symbol_keyed_member_metadata(member, name) || {}
-          description = Internal::ShapeGraph.read(member, :description)
-          metadata[:description] = description unless nil.equal?(description)
-
-          { field: key, validations:, metadata:, sensitive: sensitive || false, user_facing: user_facing || false,
-            method_call: Internal::ShapeGraph.read(member, :method_call) || false }
-        end
-
-        # A shape graph that contains itself has no traversal at all: every walk over it — this one, the
-        # runtime validator's, the schema's — recurses until the stack overflows, and SystemStackError is
-        # outside StandardError, so it escapes every rescue in the framework rather than settling into a
-        # reported failure. Rejected at declaration, where it is knowable and where the author is present.
-        def _raise_cyclic_shape!(member, name)
-          via = nil.equal?(member) ? "" : " reached from shape member #{_describe_shape_member(member, name)}"
-          raise ArgumentError,
-                "a `shape:` graph cannot contain itself — the nested shape#{via} is the same shape it is nested " \
-                "inside, so validating or reflecting it would recurse until the stack overflows. Give the nested " \
-                "shape its own members rather than reusing the shape (or the member) that encloses it."
-        end
-
-        # The generative counterpart: no object repeats, so nothing identifies a loop, and the graph is
-        # simply endless. Capped rather than walked to exhaustion, for the same reason a cycle is rejected
-        # — the alternative outcome is a SystemStackError raised while the class is being defined, which
-        # no rescue in the framework can settle.
-        def _raise_shape_too_deep!(member, name)
-          via = nil.equal?(member) ? "" : " at shape member #{_describe_shape_member(member, name)}"
-          raise ArgumentError,
-                "a `shape:` graph nested more than #{Internal::ShapeGraph::MAX_NESTING} levels deep#{via} is almost certainly " \
-                "generated rather than declared — no hand-written shape block reaches that depth, while a shape " \
-                "object that builds a fresh nested shape on every read is endless and would recurse until the " \
-                "stack overflows. Have the shape return the same finite nested shape each time it is read, or " \
-                "flatten the nesting."
-        end
-
-        # A member's declared validations, symbol-canonical at BOTH levels its grammar has: the validator names,
-        # and each validator's own option bag (see _symbolize_option_bags!, which does the same for a field).
-        # A raw `shape:` member bypasses `expects`' option handling entirely, so this is the one place its
-        # grammar gets canonicalized — and it must, because what the snapshot stores is a plain Hash: a member
-        # declared with String keys otherwise validated nothing at all, silently, and reflected an empty
-        # constraint beside it.
-        def _symbol_keyed_member_validations(member, name)
-          validations = Internal::ShapeGraph.hash_or_nil(Internal::ShapeGraph.read(member, :validations))
-          return nil if nil.equal?(validations)
-
-          _reject_defaulting_option_container!(validations) { "the validations of #{_member_owner_label(member, name)}" }
-          # A Hash axn owns, ALWAYS — canonicalized and copied in one pass. "Needs no key change" and "needs no
-          # copy" are different questions, and answering only the first left a member's options aliased to the
-          # objects the caller still held while a top-level field's were detached: mutating an `inclusion:` list
-          # afterwards widened a declared member's enum. The copy is also what the snapshot stores, so it is one
-          # allocation rather than two.
-          copy = {}
-          Internal::ShapeGraph.each_entry(validations) do |key, value|
-            canonical = case key
-                        when ::String then key.to_sym
-                        else key
-                        end
-            _raise_ambiguous_option_key!("the validations of #{_member_owner_label(member, name)}", canonical) if copy.key?(canonical)
-
-            copy[canonical] = value
-          end
-          # Each bag's own keys, then the containers themselves — through the same two helpers a field's options
-          # go through, in the same order, so a member is held to exactly what a field is held to: an
-          # `inclusion:` list keeps its class, and a container that answers with code of its own is refused
-          # unless it is frozen.
-          _symbolize_option_bags!(copy)
-          _detach_option_containers!(copy)
-          copy
-        end
-
-        def _member_owner_label(member, name) = "shape member #{_describe_shape_member(member, name)}"
-
-        # Metadata is one level of grammar (`description:` and whatever an extension registered), read as
-        # Symbols — `ShapeConfig#description` is `metadata[:description]` — so a String-keyed metadata Hash
-        # silently loses every entry it holds.
-        def _symbol_keyed_member_metadata(member, name)
-          metadata = Internal::ShapeGraph.hash_or_nil(Internal::ShapeGraph.read(member, :metadata))
-          return nil if nil.equal?(metadata)
-
-          _reject_defaulting_option_container!(metadata) { "the metadata of #{_member_owner_label(member, name)}" }
-          # Canonicalized WHILE being copied, in one pass: metadata is copied either way (what is stored IS the
-          # contract — see `_snapshot_member_attributes!`), so asking about its keys separately would be a second
-          # pass over every member for nothing.
-          copy = {}
-          Internal::ShapeGraph.each_entry(metadata) do |key, value|
-            canonical = case key
-                        when ::String then key.to_sym
-                        else key
-                        end
-            _raise_ambiguous_option_key!("the metadata of #{_member_owner_label(member, name)}", canonical) if copy.key?(canonical)
-
-            copy[canonical] = value
-          end
-          copy
-        end
-
-        # Named by class, since it has no name — through `_describe_shape_member`, so nothing of the member's
-        # own runs while the declaration error is being built.
-        def _raise_nameless_member!(member, name)
-          raise ArgumentError,
-                "a shape member must answer to `field`, naming the key it validates — the member " \
-                "#{_describe_shape_member(member, name)} answers to none. Runtime validation reads " \
-                "`member.field` for every member, so such a member would validate nothing, be omitted from " \
-                "the reflected schema entirely, and raise NoMethodError on the first call. Give it a `field` " \
-                "reader (with `validations`, the rest of the member contract), or declare it with " \
-                "`field :name` inside a `shape` block."
-        end
-
-        # A shape is a container plus the members that describe what is inside it, so a raw one that names no
-        # members list at all is malformed — it declares a shaped field nothing describes. It used to be caught
-        # on the first CALL (`ShapeValidator#check_validity!` refuses a nil members list); rejecting it here is
-        # strictly earlier and is where every other malformed declaration is answered.
-        #
-        # An empty list is NOT this: `members: []` is a real declaration (the container type still constrains the
-        # value), pointless rather than wrong, and axn's business is not to refuse it.
-        def _raise_missing_shape_members!(member, name)
-          via = nil.equal?(member) ? "" : " at shape member #{_describe_shape_member(member, name)}"
-          raise ArgumentError,
-                "a raw `shape:`#{via} must supply `members:` — a shape describes what is inside a container, so " \
-                "one with no members list constrains nothing beyond the container type, and runtime validation " \
-                "has nothing to validate against. Supply `members: [...]` (an empty list is accepted, if " \
-                "pointless), or declare the shape with a `do … end` block, which builds the members list for you."
-        end
-
-        # The other half of the documented member contract (see _raise_nameless_member!): runtime validation
-        # reads `member.validations` for every member it validates — repeatedly, and dispatched directly — so a
-        # member that answers to `field` but not `validations` declared cleanly and then raised NoMethodError on
-        # the first call. `validations: {}` is the honest spelling of "constrains nothing".
-        def _raise_member_without_validations!(member, name)
-          raise ArgumentError,
-                "a shape member must answer to `validations` as well as `field` — the member " \
-                "#{_describe_shape_member(member, name)} answers to `field` only. Runtime validation reads " \
-                "`member.validations` for every member, so such a member would raise NoMethodError on the first " \
-                "call. Give it a `validations` reader (`{}` when it constrains nothing), or declare the member " \
-                "with `field :name` inside a `shape` block."
-        end
-
-        # The same member key declared twice in one block. No comparison of the two names is needed — and so
-        # none is made: they arrived under one `to_sym` key, which is the identity the schema itself uses, so
-        # nothing a name's class can define (an `==` that raises) is dispatched to reach this conclusion.
-        def _raise_duplicate_member!(offending)
-          raise Axn::DuplicateFieldError,
-                "Duplicate shape member declared: #{_inspect_field_name(offending)} — two members of one shape would " \
-                "validate the same key, and the reflected schema would keep only the last. Declare each member once."
-        end
-
-        # Whether two colliding names are the same SPELLING — decided while a failure is already being
-        # reported, so without dispatching `==` (or anything else) on a name. A member name may be a
-        # caller-supplied String subclass, and its `==` can raise in place of the duplicate error being
-        # built, escaping every rescue when what it raises is outside StandardError.
-        #
-        # Dispatch on the shape's container — the value must match it, or it's malformed (and reaches
-        # logging before validation rejects it, so its arbitrary contents could leak). An `Array` shape
-        # maps each element (member-bearing); a `Hash` shape filters the Hash's member keys; a class
-        # (Data/Struct/PORO) shape reads members off an object ParameterFilter can't descend into. Any
-        # value whose type doesn't match the container is masked wholesale rather than treated as a lone
-        # valid element/Hash — only declared member keys would be filtered, leaking arbitrary siblings.
-        # `nil` (valid absent data) is preserved throughout via `_mask_opaque_or_preserve`.
-        #
-        # The container is compared by IDENTITY rather than with `==`: a raw `shape:` kwarg may supply any
-        # object there, and `container == Array` dispatches that object's own `==` — one answering true for
-        # both arms, or raising, would decide which redaction path a secret takes. `Array`/`Hash` are the
-        # receivers, so only their own `equal?` runs.
-        def _mask_shape_value(value, shape, action_instance)
-          container = shape[:container]
-          if ::Array.equal?(container)
-            return value.map { |element| _mask_shape_element(element, shape, action_instance) } if value.is_a?(Array)
-
-            return _mask_opaque_or_preserve(value)
-          end
-          return _mask_shape_element(value, shape, action_instance) if ::Hash.equal?(container)
-
-          _mask_opaque_or_preserve(value)
-        end
-
-        # A non-Hash value where members are expected is opaque to ParameterFilter → mask it whole. A Hash
-        # keeps its own keys (ParameterFilter redacts the sensitive ones); only recurse into a nested-shape
-        # member's value when that nested shape actually carries a sensitive member, to avoid needless
-        # masking of an unrelated non-Hash deeper down. The member key is matched in either symbol or
-        # string form, since extraction accepts both.
-        def _mask_shape_element(element, shape, action_instance)
-          return _mask_opaque_or_preserve(element) unless element.is_a?(Hash)
-
-          _sensitive_nested_members(shape, action_instance).each_with_object(element.dup) do |(member, nested), masked|
-            _present_key_variants(masked, member.field).each do |key|
-              masked[key] = _mask_shape_value(masked[key], nested, action_instance)
-            end
-          end
-        end
-
-        # The `[(member, nested_shape)]` pairs a mask has to descend into: the members of `shape` whose OWN
-        # nested shape carries a sensitive member. Memoized per shape (by identity) on the same condition as
-        # `_sensitive_shape_paths` — no `sensitive:` resolving against the action — because for the ordinary
-        # flat shape the answer is EMPTY, and finding that out cost a `nested_shape` read (a bound-Method
-        # allocation each) for every member of the shape, on every value masked.
-        def _sensitive_nested_members(shape, action_instance)
-          return _derive_sensitive_nested_members(shape, action_instance) if _has_dynamic_sensitive_fields?
-
-          _contract_redaction.nested_members_for(shape) { _derive_sensitive_nested_members(shape, nil) }
-        end
-
-        def _derive_sensitive_nested_members(shape, action_instance)
-          Internal::ShapeGraph.members(shape).filter_map do |member|
-            nested = _member_shape(member)
-            next unless nested && _shape_has_sensitive_member?(nested, action_instance)
-
-            [member, nested]
-          end
-        end
-
-        # The names of the `sensitive:` members reachable inside one shape-bearing config's value — the flat
-        # key set `inspect` hands to a `ParameterFilter`, since a member redacts by NAME wherever it appears
-        # (every array element, any nesting depth), unlike the precise wire path a subfield contributes.
-        # Memoized on the same condition as `_sensitive_shape_paths` — no `sensitive:` resolving against the
-        # action — since `inspect` asks it once per displayed field, and the walk covers the whole stored graph.
-        #
-        # Lives here rather than with the inspector because it is a question about the CONTRACT, and because
-        # sensitivity is read through `_config_sensitive` — the same seam logging reads — so a member that
-        # defines `sensitive:` cannot escape redaction in `inspect` by denying the reader while logging
-        # redacts it anyway.
-        def _sensitive_member_names(config, action_instance)
-          return _derive_sensitive_member_names(config, action_instance) if _has_dynamic_sensitive_fields?
-
-          _contract_redaction.member_names_for(config) { _derive_sensitive_member_names(config, nil) }
-        end
-
-        # Bounded both ways, exactly as `_flatten_sensitive_candidates` is, and for its reason: a config assigned
-        # onto a class rather than declared carries a shape no walk has traversed, which can be pointed at itself
-        # (which `CycleGuard` sees) or mint a fresh one on every read (which nothing but depth sees). `inspect` is
-        # reachable directly, not only from a side channel, so an unbounded walk here raises SystemStackError at
-        # the caller rather than degrading a log line.
-        #
-        # A cycle re-reaches members an enclosing frame is already collecting, so stopping loses nothing.
-        # Past the depth bound nothing can enumerate a graph that mints its members on demand — and the
-        # value it describes is masked wholesale by then anyway (`_shape_has_sensitive_member?` answers true
-        # past the same bound), so `inspect` shows a redacted value rather than a leaked one.
-        def _derive_sensitive_member_names(config, action_instance, seen = nil, depth = 0)
-          shape = Internal::ShapeGraph.shape_in(config.validations)
-          return [] if nil.equal?(shape) || depth > Internal::ShapeGraph::MAX_NESTING
-
-          Axn::Internal::CycleGuard.guard(shape, seen, on_cycle: []) do |open|
-            # Through the shared seam, for the reason _flatten_sensitive_candidates gives: a list hiding
-            # itself from `flat_map` would drop a sensitive member from the redaction set.
-            Internal::ShapeGraph.members(shape).flat_map do |member|
-              names = _derive_sensitive_member_names(member, action_instance, open, depth + 1)
-              names << member.field if _member_sensitive?(member, action_instance)
-              names
-            end
-          end
-        end
-
-        # Private, though declared next to their definitions rather than relocated below: these are
-        # declaration-time internals of the shape-member walk, called only with an implicit receiver (from here
-        # and from ContractForSubfields, which is extended onto the same class). Kept in place so each stays
-        # beside the walk it belongs to, with its comment.
-        #
-        # Only these — the ones this PR added. The other leading-underscore public class methods are
-        # `class_attribute` accessors and long-standing declaration hooks the framework and downstream gems
-        # already reach; narrowing those is a separate, breaking question.
-        private :_spend_paths!, :_raise_too_many_member_paths!, :_symbol_keyed_member_validations,
-                :_symbol_keyed_member_metadata, :_snapshot_member_attributes!,
-                :_member_owner_label, :_describe_shape_member,
-                :_snapshot_declared_shape!, :_validate_and_snapshot_shape!, :_walk_shape_graph!,
-                :_check_and_copy_shape_members!, :_raise_cyclic_shape!, :_raise_shape_too_deep!,
-                :_raise_duplicate_member!, :_raise_nameless_member!,
-                :_raise_missing_shape_members!, :_raise_member_without_validations!
 
         private
 
@@ -1576,10 +677,6 @@ module Axn
         # `_build_shape_member` rejects them explicitly with the reader-less reason instead.
         SHAPE_MEMBER_READER_OPTIONS = %i[as prefix].freeze
 
-        # The mask a sensitive value is replaced with — matches `ActiveSupport::ParameterFilter`'s default
-        # so wholesale-masked values read identically to per-key-filtered ones.
-        SENSITIVE_FILTERED_MASK = "[FILTERED]"
-
         # Parse a structured field's block into a `{ members: [...], container: <klass> }` validation
         # value. `container` lets ShapeValidator defer a type mismatch to TypeValidator (rather than
         # trying to extract members from the wrong kind of value).
@@ -1700,142 +797,6 @@ module Axn
         # method a walk requires. Assembled as a plain Hash rather than taken from the shape's `merge`,
         # which a subclass can override to return anything — including a non-Hash, or something that drops
         # the container just derived.
-        # A contract must not change after the class is declared, and an option value the caller still holds is
-        # aliased into it: mutating a `validate:` bag swapped the validator a declared field runs, a mutated `of:`
-        # bag changed a declared element type, and appending to an `inclusion:` list widened a declared enum —
-        # each after the fact, on a class already defined.
-        #
-        # Detached one level deep, which is exactly the boundary that matters: the plain Hash/Array CONTAINERS
-        # axn stores are copied, while the values inside them stay the caller's objects — a `validate:` callable,
-        # a `model:` class, an `inclusion:` member. Those are meant to be the caller's, and copying them would
-        # change what a declaration means rather than protect it. `shape:` is excluded because it needs a deep
-        # copy of its own (see _validate_and_snapshot_shape!) and gets one downstream.
-        # Nothing an option container can define decides whether it is detached, or what the detached copy holds.
-        # The type tests are `case`/`when` (`Module#===`, a C-level check) rather than `is_a?`, and the copies are
-        # taken through bound primitives (see ShapeGraph) rather than the container's own `transform_values`/`dup`.
-        # A subclass answering `is_a?(Array)` with false, or whose `dup` returned `self`, or whose
-        # `transform_values` handed back the receiver, otherwise stayed aliased into the declared contract while
-        # the plain-Array case beside it was correctly copied. An ARRAY that owns any of those is now refused
-        # before the copy is attempted (see _detached_option_array), so the bound `dup` is belt-and-braces there;
-        # a BAG is not, since it is copied entry-wise whatever its class, which is what the bound `Hash#each`
-        # holds.
-        #
-        # An Array keeps its CLASS (a same-class `dup`, or the caller's own object when it is already frozen —
-        # see _detached_option_array) because its own class is part of what a declaration means — a frozen
-        # `inclusion:` set answers membership with its own `include?`, and reflection withholds an enum for
-        # anything but an exact Array. A bag becomes a plain Hash — which is what it already became, and axn
-        # reads bags with `[]`/`dig` only.
-        def _detach_option_containers!(validations)
-          validations.each do |key, value|
-            next if key == :shape
-
-            case value
-            when ::Hash then validations[key] = _detached_option_bag(key, value)
-            when ::Array then validations[key] = _detached_option_array(value, "`#{key}:`")
-            end
-          end
-        end
-
-        def _detached_option_bag(key, bag)
-          _reject_defaulting_option_container!(bag) { "the `#{key}:` option bag" }
-          copy = Internal::ShapeGraph.copy_entries(bag)
-          copy.each do |option_key, option|
-            case option
-            when ::Array then copy[option_key] = _detached_option_array(option, "`#{key}: { #{option_key}: … }`")
-            end
-          end
-          copy
-        end
-
-        # A caller Hash that answers missing keys from a DEFAULT is refused wherever a declaration would store one.
-        # It is the same split the copy exists to close, arriving from the other side: axn copies every container
-        # it stores entry-wise, and a default is not an entry, so the options such a Hash answers are simply not
-        # in the stored contract — while a guard reading the original with `[]` sees them. The author is told
-        # rather than left to find out, per the option-key rule above: an option is never silently ignored.
-        #
-        # Refusing rather than carrying the default over, because carrying it cannot make the declaration work.
-        # ActiveModel builds each validator's options into a Hash of its OWN (`_parse_validates_options`), so a
-        # default never reached a validator even when axn stored the caller's bag: `expects :a, type:
-        # Hash.new(String)` raised "must supply :klass" on every CALL — a key the author believes they supplied —
-        # before this copy existed and after it. Declaration is where that is knowable.
-        #
-        # The label is YIELDED, so naming the container costs nothing until there is an error to name (see
-        # `_symbol_keyed_bag`), and the two readers consulted are Hash's own (see ShapeGraph.supplies_default?).
-        def _reject_defaulting_option_container!(hash)
-          return unless Internal::ShapeGraph.supplies_default?(hash)
-
-          raise ArgumentError,
-                "#{yield} answers a missing key from a Hash default (`Hash.new(…)` or a `default_proc`) rather " \
-                "than from an entry of its own, and axn cannot carry that into the contract: a declared " \
-                "container is copied entry-wise so that mutating what you still hold cannot change an " \
-                "already-declared class, and ActiveModel rebuilds a validator's options into a Hash of its own " \
-                "besides — so an option supplied through the default is dropped, and the declaration fails on a " \
-                "call complaining about a key you did supply. Write the options out as entries " \
-                "(`type: { klass: String }`)."
-        end
-
-        # Three outcomes, and which one a container gets is decided WITHOUT running any of its code.
-        #
-        # A container axn can copy faithfully by construction is copied, and the copy is the contract. That
-        # condition is that the container answers NOTHING with code of its own (`NativeMethods.own_array_methods`
-        # is empty): `Kernel#dup` copies the elements, so the copy answers as the original exactly where every
-        # answer is a pure function of the elements, which is exactly where every answer is Ruby's own. Every
-        # plain Array passes, as does a subclass that adds no methods. A FROZEN container is stored as it is,
-        # whatever it owns: the copy exists so that mutating what the caller still holds cannot change a declared
-        # contract, and a frozen container cannot be mutated, so there is nothing to detach it from. Anything
-        # else is REFUSED at declaration.
-        #
-        # The refusal is deliberate over-rejection, and it replaces two rounds of verifying the copy and one of
-        # gating on the duplication hooks. Comparing the copy's ELEMENTS missed a hook that dropped a derived
-        # lookup index the container's own `include?` reads: the elements survived and the copy rejected every one
-        # of them. Asking the copy `include?` about each element then missed a hook that dropped only the index of
-        # accepted NON-elements — a set holding `"canon"` and aliasing `"alias"` to it accepted both, its copy
-        # accepted only `"canon"`, and no element-based probe can see a difference outside the elements. Gating on
-        # the duplication HOOKS then missed the copy's other two differences from the original: `dup` shares the
-        # instance variables and drops the singleton class, so a membership derived from `self`, from an ivar, or
-        # from a singleton method diverges with entirely native duplication (and an ivar-derived one is still the
-        # caller's to mutate afterwards, which is the aliasing the copy exists to prevent). Ownership of
-        # everything the container answers with is a fact rather than a prediction (see Internal::NativeMethods),
-        # and a container that would copy faithfully is over-rejected by it with a bounded way to stay legal:
-        # freeze it.
-        #
-        # Two residues, stated rather than papered over, and both are the same one-level depth the copy promises.
-        # A FROZEN container's elements — and whatever its ivars point at — are still the caller's objects, so a
-        # frozen container deriving membership from a mutable index can still be widened after the class is
-        # declared: freezing promises that axn stores what you froze, and how deep that goes is the author's.
-        # And a membership container that is not an Array (a `Set`, a `Range`, an object answering `include?`) is
-        # not reached here at all: it is stored as the caller's object, so nothing of axn's answers membership and
-        # there is nothing to diverge — but nothing detaches it either. A `Range` is frozen by construction; a
-        # `Set` is mutable. Both residues are recorded in `property_name_collision_spec.rb`.
-        #
-        # The container is named by class and its methods by the method table, never by `inspect` — its own code
-        # must not run while the declaration error it caused is being built, and a method name is rendered like
-        # any other name reaching prose (a non-UTF-8 one would otherwise raise while the error is built).
-        def _detached_option_array(value, label)
-          return value if Internal::NativeMethods.frozen?(value)
-
-          own = Internal::NativeMethods.own_array_methods(value)
-          return Internal::ShapeGraph.detached_dup(value) if own.empty?
-
-          raise ArgumentError,
-                "the #{label} container (of class #{Axn::Internal::ClassName.of(value)}) defines methods of its " \
-                "own (#{_describe_own_methods(own)}), so axn cannot copy it. A declared contract is copied at " \
-                "declaration so that mutating what you still hold cannot change it — and `dup` copies the " \
-                "elements while sharing the instance variables and dropping the singleton class, so the copy " \
-                "answers as you declared only where the answer is Ruby's own. What your code answers in the copy " \
-                "cannot be established without running it, so a copy that silently rejects the values you " \
-                "declared is indistinguishable from a faithful one. Supply a plain Array, or freeze this " \
-                "container (a frozen one is stored as-is, since nothing can mutate it afterwards)."
-        end
-
-        # The first few offending names, so the author is pointed at the method to move or the object to freeze
-        # rather than at a rule. Sorted for a stable message, and capped because a rich subclass has dozens.
-        def _describe_own_methods(names)
-          shown = names.uniq.sort
-          rendered = shown.first(3).map { |name| "`#{_inspect_field_name(name)}`" }.join(", ")
-          shown.size > 3 ? "#{rendered}, and #{shown.size - 3} more" : rendered
-        end
-
         def _derive_raw_shape_container!(validations)
           shape = Internal::ShapeGraph.hash_or_nil(validations[:shape])
           return if nil.equal?(shape)
@@ -2204,7 +1165,7 @@ module Axn
           allow_blank: false,
           **validations
         )
-          _detach_option_containers!(validations)
+          Internal::ShapeGraph.detach_option_containers!(validations)
           _canonicalize_blank_gates!(validations)
 
           # `coerce: <Type>` sugar → a coerce flag inside the type bag (coercion binds to the type;
